requirements.txt
lightning
torch
wandb
numpy
pandas
scipy
python-dotenv
tqdm
yaml
soundfile
fairseq
